% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gride_autoplots.R
\name{autoplot.byf_gride}
\alias{autoplot.byf_gride}
\title{Plot the posterior distribution for the Gride model}
\usage{
\method{autoplot}{byf_gride}(object, chain = F, ...)
}
\arguments{
\item{object}{an object of class \code{byf_gride}, the output of the function \code{bayesfit_gride}.}

\item{chain}{logical. If \code{TRUE}, the MCMC chain is reported. Otherwise, it displays the posterior density.}

\item{...}{other arguments passed to specific methods.}
}
\description{
Plot the posterior distribution for the Gride model
}
