% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoNN_autoplot.R
\name{autoplot.byf_twonn}
\alias{autoplot.byf_twonn}
\title{Plot the TWO-NN posterior distribution}
\usage{
\method{autoplot}{byf_twonn}(object, plot_low = 0, plot_upp = NULL, by = 0.05, ...)
}
\arguments{
\item{object}{an object of class \code{byf_twonn}, the output of the \code{bayesfit_twonn} function.}

\item{plot_low}{the lower bound of the interval on which the posterior density is plotted.}

\item{plot_upp}{the upper bound of the interval on which the posterior density is plotted.}

\item{by}{the step-size at which the sequence spanning the interval is incremented}

\item{...}{other arguments passed to specific methods.}
}
\value{
a plot of the posterior distribution for the TWO-NN model
}
\description{
Plot the TWO-NN posterior distribution
}
\examples{
\dontrun{
m <- bayesfit_twonn(data)
plot(m)
}
}
