% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gride_evolution.R
\name{autoplot.gride_evolution}
\alias{autoplot.gride_evolution}
\title{Plot the evolution of \code{Gride} estimates}
\usage{
autoplot.gride_evolution(object, title = "Gride Evolution", ...)
}
\arguments{
\item{object}{an object of class \code{gride_evolution}.}

\item{title}{an optional string to customize the title of the plot.}

\item{...}{other arguments passed to specific methods.}
}
\value{
object of class \code{\link[ggplot2]{ggplot}}. It displays the
the evolution of the Gride maximum likelihood estimates as a function
of the average distance from \code{n2}.
}
\description{
Use this method without the \code{.gride_evolution} suffix and after loading
the \code{ggplot2} package. It plots the evolution of the \code{id}
estimates as a function of the average distance from the furthest NN of
each point.
}
