% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot_gride.R
\name{autoplot.gride_mle}
\alias{autoplot.gride_mle}
\title{Plot the simulated bootstrap sample for the MLE \code{Gride}}
\usage{
\method{autoplot}{gride_mle}(object, title = "MLE Gride - Bootstrap sample", ...)
}
\arguments{
\item{object}{object of class \code{gride_mle}.
It is obtained using the output of the \code{gride} function when
\code{method = "mle"}.}

\item{title}{title for the plot.}

\item{...}{other arguments passed to specific methods.}
}
\value{
object of class \code{\link[ggplot2]{ggplot}}. It displays the
density plot of the sample generated via parametric bootstrap to help the
visual assessment of the uncertainty of the \code{id} estimates.
}
\description{
Use this method without the \code{.gride_mle} suffix.
It display the density plot of sample obtained via
parametric bootstrap for the \code{Gride} model.
}
\seealso{
\code{\link{gride}}
}
