% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hidalgo_autoplots.R
\name{autoplot.hid_all_mcmc}
\alias{autoplot.hid_all_mcmc}
\title{Plot the postprocessed MCMC trace plots}
\usage{
\method{autoplot}{hid_all_mcmc}(object, ...)
}
\arguments{
\item{object}{an object of class \code{hid_all_mcmc}, the output of the \code{Hidalgo_postpr_chains} function when all_chains is \code{TRUE}.}

\item{...}{other arguments passed to specific methods.}
}
\value{
the trace plots of the intrinsic dimension  each observation after the postprocessing
}
\description{
Plot the postprocessed MCMC trace plots
}
