% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hidalgo_autoplots.R
\name{autoplot.hid_class}
\alias{autoplot.hid_class}
\title{Plot posterior \code{id} for each observation stratified by an external factor}
\usage{
\method{autoplot}{hid_class}(object, class, type = "histogram", ...)
}
\arguments{
\item{object}{an object of class \code{hid_class}, the output of the \code{Hidalgo_ID_class} function.}

\item{class}{a factor according to the observations can be stratified by.}

\item{type}{character, provides the type of the plot. It must be one of the following options: histogram, density, boxplot, or violin plot.}

\item{...}{other arguments passed to specific methods.}
}
\value{
a plot of the \code{id} estimates stratified by class.
}
\description{
Plot posterior \code{id} for each observation stratified by an external factor
}
