% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hidalgo_autoplots.R
\name{autoplot.hid_psm}
\alias{autoplot.hid_psm}
\title{Plot the Posterior Coclustering Matrix}
\usage{
\method{autoplot}{hid_psm}(object, id_names = NULL, class = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{hid_psm}, the output of the function \code{Hidalgo_coclustering_matrix}.}

\item{id_names}{a vector of label identifying each observation.}

\item{class}{factor according to the observations can be stratified by.}

\item{...}{other arguments passed to specific methods.}
}
\value{
Heatmap displaying the Posterior Coclustering Matrix. Columns and rows can be ordered by an external class and/or the optimal partition
}
\description{
Plot the Posterior Coclustering Matrix
}
\examples{
\dontrun{
Hidalgo_coclustering_matrix(output, class = group_variable, VI = F, plot = F)
}
}
