% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hidalgo_autoplots.R
\name{autoplot.hid_sum_mcmc}
\alias{autoplot.hid_sum_mcmc}
\title{Plot mean and median of the \code{id} for each observation after postprocessing}
\usage{
\method{autoplot}{hid_sum_mcmc}(object, ...)
}
\arguments{
\item{object}{an object of class \code{hid_sum_mcmc}, the output of the \code{Hidalgo_postpr_chains} function when all_chains is \code{FALSE}.}

\item{...}{other arguments passed to specific methods.}
}
\value{
a plot displaying median and mean for each observations computed over the postprocessed \code{id} chains.
}
\description{
Plot mean and median of the \code{id} for each observation after postprocessing
}
