% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hidalgo_autoplots.R
\name{autoplot.Hidalgo}
\alias{autoplot.Hidalgo}
\title{Plot the intrinsic dimension chains}
\usage{
\method{autoplot}{Hidalgo}(object, ...)
}
\arguments{
\item{object}{an object of class \code{Hidalgo}, the output of the \code{Hidalgo} function.}

\item{...}{other arguments passed to specific methods.}
}
\value{
Plots the MCMC of the different intrinsic dimensions.
}
\description{
Plot the intrinsic dimension chains
}
