% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoNN_autoplot.R
\name{autoplot.lnf_twonn}
\alias{autoplot.lnf_twonn}
\title{Plot a summary of the Linear Fit TWO-NN model}
\usage{
\method{autoplot}{lnf_twonn}(object, ...)
}
\arguments{
\item{object}{an object of class \code{lnf_twonn}, the output of the \code{linfit_twonn} function.}

\item{...}{other arguments passed to specific methods.}
}
\value{
a plot that shows the goodness of the linear fit
}
\description{
Plot a summary of the Linear Fit TWO-NN model
}
\examples{
\dontrun{
m <- linfit_twonn(data)
plot(m)
}
}
