% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gride_autoplots.R
\name{autoplot.mcmc_gride}
\alias{autoplot.mcmc_gride}
\title{Plot the simulated MCMC chains}
\usage{
\method{autoplot}{mcmc_gride}(object, ...)
}
\arguments{
\item{object}{an object of class \code{autoplot.mcmc_gride}, the output of the \code{posterior_sampler_gride} function.}

\item{...}{other arguments passed to specific methods.}
}
\value{
the trace plots of the posterior simulations for the Gride model
}
\description{
Plot the simulated MCMC chains
}
