% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twonn_autoplot.R
\name{autoplot.twonn_mle}
\alias{autoplot.twonn_mle}
\title{Plot the output of the \code{TWO-NN} model estimated via the Maximum
Likelihood approach}
\usage{
autoplot.twonn_mle(object, title = "MLE TWO-NN", ...)
}
\arguments{
\item{object}{object of class \code{twonn_mle}, the output of the
\code{twonn} function when \code{method = "mle"}.}

\item{title}{character string used as title of the plot.}

\item{...}{other arguments passed to specific methods.}
}
\value{
\code{\link[ggplot2]{ggplot2}} object displaying the point estimate
and confidence interval obtained via maximum likelihood approach of the
\code{id} parameter.
}
\description{
Use this method without the \code{.twonn_mle} suffix and after loading the
\code{ggplot2} package.
The function returns the point estimate along with the confidence bands
computed via the \code{mle} method.
}
\seealso{
\code{\link{twonn}}
}
