% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gride.R
\name{bayesfit_gride}
\alias{bayesfit_gride}
\title{Gride estimator for ID under the Bayesian framework}
\usage{
bayesfit_gride(
  X,
  DM = NULL,
  n1,
  n2,
  sigma = 0.5,
  a_d = 1,
  b_d = 1,
  alpha = 0.95,
  nsim = 5000,
  burn_in = 2000
)
}
\arguments{
\item{X}{a \code{n} \eqn{times} \code{D} data matrix.}

\item{DM}{a distance matrix describing the nearest neighbors structure. Default is \code{NULL}.}

\item{n1}{order of the NN according to the first distance is computed.}

\item{n2}{order of the NN according to the second distance is computed.}

\item{sigma}{standard deviation of the Gaussian proposal used in the MH step.}

\item{a_d}{shape parameter of the Gamma prior distribution for \eqn{d}.}

\item{b_d}{rate parameter of the Gamma prior distribution for \eqn{d}.}

\item{alpha}{the amount of probability contained in the posterior credible interval.}

\item{nsim}{the number of MCMC iterations to collect in the sample.}

\item{burn_in}{the number of iterations to discard from the MCMC sample.}
}
\value{
the vector of summary statistics from the posterior MCMC sample.
}
\description{
Gride estimator for ID under the Bayesian framework
}
\examples{
\dontrun{
bayesfit_gride(X=dataset, n1 = 4, n2 = 8)
}
}
