% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoNN.R
\name{bayesfit_twonn}
\alias{bayesfit_twonn}
\title{Bayesian estimator for the Instrisic Dimension}
\usage{
bayesfit_twonn(
  X,
  a_d = 0.001,
  b_d = 0.001,
  alpha = 0.95,
  trimmed = F,
  DM = NULL,
  alpha_trimmed = 0.01
)
}
\arguments{
\item{X}{a \code{n} \eqn{times} \code{D} data matrix.}

\item{a_d}{shape parameter of the Gamma prior on \eqn{d}.}

\item{b_d}{rate parameter of the Gamma prior on \eqn{d}.}

\item{alpha}{the order of the quantile for the credible set.}

\item{trimmed}{logical, if \code{TRUE}, the trimming of the most extreme observations is performed.}

\item{DM}{a distance matrix describing the nearest neighbors structure. Default is \code{NULL}.}

\item{alpha_trimmed}{proportion of trimmed observations.}
}
\value{
quantiles, mean, mode and median of the posterior distribution on \eqn{d}.
}
\description{
Bayesian estimator for the Instrisic Dimension
}
