% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoNN_Gride_Aux.R
\name{beta_prime}
\alias{beta_prime}
\title{Random sampler for observations generated from a Beta prime distribution}
\usage{
beta_prime(nsim, a, b)
}
\arguments{
\item{nsim}{integer, the number of observations to generate.}

\item{a}{the first shape parameter of the distribution.}

\item{b}{the second shape parameter of the distribution.}
}
\value{
a vector of length \eqn{nsim} sampled from a Beta prime distribution.
}
\description{
Random sampler for observations generated from a Beta prime distribution
}
\examples{
Y <- beta_prime(100,3,2)

}
