% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gride.R
\name{bootstrap_gride}
\alias{bootstrap_gride}
\title{Bootstrap sample generator for the Gride maximum likelihood estimate of the intrinsic dimension}
\usage{
bootstrap_gride(X, n1, n2, nsim = 2000)
}
\arguments{
\item{X}{a \code{n} \eqn{times} \code{D} data matrix.}

\item{n1}{order of the NN according to the first distance is computed.}

\item{n2}{order of the NN according to the second distance is computed.}

\item{nsim}{integer,the number of bootstrap replications to consider.}
}
\value{
a matrix of dimension \code{n} \eqn{\times} \code{nsim} with the bootstrap samples organized by column.
}
\description{
Bootstrap sample generator for the Gride maximum likelihood estimate of the intrinsic dimension
}
\examples{
\dontrun{
Bootstrap_gride(Data,2,5)
}
}
