% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gride.R
\name{d_mudot}
\alias{d_mudot}
\title{Density for the generic ratio of two NNs' distances}
\usage{
d_mudot(mu.dot, n1, n2, d, log = F)
}
\arguments{
\item{mu.dot}{real positive value, the quantile to be considered.}

\item{n1}{order of the NN according to the first distance is computed.}

\item{n2}{order of the NN according to the second distance is computed.}

\item{d}{real value, the ID parameter.}

\item{log}{logical, if \cite{TRUE}, the log-density is returned.}
}
\value{
the value of the (log-)density in the specified quantile.
}
\description{
Density for the generic ratio of two NNs' distances
}
\examples{
\dontrun{
d_mudot(x=4, n1 = 1, n2 = 2, d=4)
}

}
