% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoNN_Gride_Aux.R
\name{generate_mudots}
\alias{generate_mudots}
\title{Ratio statistics \eqn{\dot{\mu}} needed for the Intrinsic Dimension estimation}
\usage{
generate_mudots(X, n1, n2, dupl_remove = T, DM = NULL)
}
\arguments{
\item{X}{a \code{n} \eqn{times} \code{D} data matrix.}

\item{n1}{order of the NN according to the first distance is computed.}

\item{n2}{order of the NN according to the second distance is computed.}

\item{dupl_remove}{logical, should duplicates in the data be removed automatically?}

\item{DM}{a distance matrix describing the nearest neighbors structure. Default is \code{NULL}.}
}
\value{
a vector of length \eqn{n}.
}
\description{
Ratio statistics \eqn{\dot{\mu}} needed for the Intrinsic Dimension estimation
}
\examples{
\dontrun{
generate_mudots(dataset, n1 = 1, n2 = 2)
}
}
