% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoNN_Gride_Aux.R
\name{generate_mus}
\alias{generate_mus}
\title{Computes the ratio statistics \eqn{\mu} needed for the Intrinsic Dimension estimation}
\usage{
generate_mus(X, dupl_remove = T, DM = NULL)
}
\arguments{
\item{X}{a \code{n} \eqn{times} \code{D} data matrix.}

\item{dupl_remove}{logical, should duplicates in the data be removed automatically?}

\item{DM}{a distance matrix describing the nearest neighbors structure. Default is \code{NULL}.}
}
\value{
a vector of length \eqn{n}.
}
\description{
Computes the ratio statistics \eqn{\mu} needed for the Intrinsic Dimension estimation
}
