% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hidalgo_MAIN.R
\name{Hidalgo}
\alias{Hidalgo}
\title{Gibbs sampler for the Hidalgo model}
\usage{
Hidalgo(
  X = NULL,
  DM = NULL,
  K = 10,
  nsim = 1000,
  burn_in = 1000,
  thinning = 1,
  verbose = T,
  q = 3,
  csi = 0.75,
  alpha_Dirichlet = 0.05,
  a0_d = 1,
  b0_d = 1,
  prior_type = "Conjugate",
  D = NULL,
  pi_mass = 0.5,
  recap = T,
  if_coda = T
)
}
\arguments{
\item{X}{\code{n} \eqn{times} \code{D} data matrix.}

\item{DM}{\code{n} \eqn{times} \code{n} distance matrix (can be alternative to \code{X}).}

\item{K}{integer, number of mixture components.}

\item{nsim}{number of MCMC iterations to run.}

\item{burn_in}{number of MCMC iterations to discard as burn-in period.}

\item{thinning}{thinning interval.}

\item{verbose}{logical, should the progress of the sampler be printed?}

\item{q}{integer, first local homogeneity parameter.}

\item{csi}{real between 0 and 1, second local homogeneity parameter.}

\item{alpha_Dirichlet}{parameter of the Dirichlet prior on the mixture weights.}

\item{a0_d}{shape parameter of the Gamma prior on \code{d}.}

\item{b0_d}{rate parameter of the Gamma prior on \code{d}.}

\item{prior_type}{character, type of Gamma prior on \code{d}, can be \code{Conjugate}, \code{Truncated}, or \code{Truncated_PointMass}.}

\item{D}{integer, the maximal dimension of the dataset.}

\item{pi_mass}{probability placed a priori on \code{D} when \code{Truncated_PointMass} is chosen.}

\item{recap}{logical, includes a list of the chosen parameters in the output.}

\item{if_coda}{logical, if \code{TRUE} the MCMC is returned as a coda object.}
}
\value{
a list containing
    \itemize{
    \item \code{membership_labels} chains of the membership labels for all the observations;
    \item \code{cluster_prob} chains of the posterior mixture weights;
    \item \code{intrinsic_dimension} chains of the K intrinsic dimensions, one per mixture component.
    }
}
\description{
Gibbs sampler for the Hidalgo model
}
