% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hidalgo_Aux.R
\name{Hidalgo_coclustering_matrix}
\alias{Hidalgo_coclustering_matrix}
\title{Compute the Posterior Coclustering Matrix}
\usage{
Hidalgo_coclustering_matrix(object, estimate_clustering = TRUE, greed = F)
}
\arguments{
\item{object}{the output of the \code{Hidalgo} function.}

\item{estimate_clustering}{logical, if \code{TRUE}, the optimal clustering based on the minimization of the Variation of Information is computed. Can be slow with more than 1000 data points.}

\item{greed}{logical, if \code{TRUE} the greedy version of the VI algorithm is used. Use this option when the dataset is in the order of the hundreds}
}
\value{
a list containing the posterior coclustering matrix and, if requested, the optimal partition.
}
\description{
Compute the Posterior Coclustering Matrix
}
\examples{
\dontrun{
Hidalgo_coclustering_matrix(output, class = group_variable, estimate_clustering = F, plot = F)
}
}
