% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hidalgo_Aux.R
\name{Hidalgo_data_preprocessing}
\alias{Hidalgo_data_preprocessing}
\title{Preprocess the data for the Hidalgo model}
\usage{
Hidalgo_data_preprocessing(X = NULL, q, DM = NULL)
}
\arguments{
\item{X}{a \code{n} \eqn{`times} \code{D} data matrix.}

\item{q}{integer, the local homogeneity parameter.}

\item{DM}{a distance matrix describing the nearest neighbors structure. Default is \code{NULL}.}
}
\value{
A list containing
      \itemize{
      \item \code{mu} the vector of the ratios NN/1NN.
      \item \code{Nq} the matrix \eqn{N^q}.
      \item \code{q} the provided local homogeneity parameter.
      }
}
\description{
Preprocess the data for the Hidalgo model
}
