% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hidalgo_Aux.R
\name{Hidalgo_ID_class}
\alias{Hidalgo_ID_class}
\title{Compute mean and median ID of the postprocessed chains stratified by an external factor}
\usage{
Hidalgo_ID_class(output, class, avg = T)
}
\arguments{
\item{output}{the output of the \code{Hidalgo} function.}

\item{class}{a factor according to the observations can be stratified by.}

\item{avg}{logical, if \code{TRUE} the posterior average is computed for each observation's chain. Otherwise, the median is used.}
}
\value{
a tibble containing the \code{id} estimates stratified by class.
}
\description{
Compute mean and median ID of the postprocessed chains stratified by an external factor
}
