% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hidalgo_Aux.R
\name{Hidalgo_postpr_chains}
\alias{Hidalgo_postpr_chains}
\title{Postprocess the \code{id} chains to disentangle the label switching}
\usage{
Hidalgo_postpr_chains(output, all_chains = F)
}
\arguments{
\item{output}{the output of the \code{Hidalgo} function.}

\item{all_chains}{: logical, if TRUE all the chains, after the disentanglement for label switching, are reported.  If \code{FALSE}, a dataset with summary statistics is returned instead.}
}
\value{
a dataset \code{nsim} \eqn{times} \code{n}, with the observation specific chain of the id.
}
\description{
Postprocess the \code{id} chains to disentangle the label switching
}
