% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{index_row_col}
\alias{index_row_col}
\title{Summarize the Nq matrix into indexes of rows and columns}
\usage{
index_row_col(Nq, q, N)
}
\arguments{
\item{Nq}{the binary adjacency matrix}

\item{q}{number of neighbors considered}

\item{N}{number of observations}
}
\description{
Summarize the Nq matrix into indexes of rows and columns
}
