% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intRinsic-package.R
\docType{package}
\name{intRinsic-package}
\alias{intRinsic}
\alias{intRinsic-package}
\title{intRinsic: Instrinsic dimension estimators based on the TwoNN and Hidalgo models}
\description{
The package contains the functions to fit the TwoNN-type models and the Hidalgo Bayesian mixture model.
}
\author{
\strong{Maintainer}: Francesco Denti \email{fdenti@uci.edu}

}
\keyword{internal}
