% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoNN.R
\name{linfit_twonn}
\alias{linfit_twonn}
\title{Least Square Estimator for the Intrinsic Dimension}
\usage{
linfit_twonn(X, trimmed = F, alpha_trimmed = 0.01, DM = NULL)
}
\arguments{
\item{X}{a n \eqn{times} D data matrix.}

\item{trimmed}{logical, if \code{TRUE}, the trimming of the most extreme observations is performed.}

\item{alpha_trimmed}{proportion of trimmed observations.}

\item{DM}{a distance matrix describing the nearest neighbors structure. Default is \code{NULL}.}
}
\value{
the estimate for the ID and, if requested, the plot depicting the linear fit.
}
\description{
Least Square Estimator for the Intrinsic Dimension
}
