% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loglikelihood_gride}
\alias{loglikelihood_gride}
\title{Log-likelihood generated from the generic ratio of two NN's distances}
\usage{
loglikelihood_gride(par, n1, n2, mu_dots)
}
\arguments{
\item{par}{the ID, variable of the objective function,}

\item{n1}{order of the NN according to the first distance is computed.}

\item{n2}{order of the NN according to the second distance is computed.}

\item{mu_dots}{the ratios of distance of generic order.}
}
\description{
Log-likelihood generated from the generic ratio of two NN's distances
}
