% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gride.R
\name{loglikelihood_gride_R}
\alias{loglikelihood_gride_R}
\title{Log-likelihood generated from the generic ratio of two NN's distances}
\usage{
loglikelihood_gride_R(par, n1, n2, mu_dots)
}
\arguments{
\item{par}{the ID, variable of the objective function,}

\item{n1}{order of the NN according to the first distance is computed.}

\item{n2}{order of the NN according to the second distance is computed.}

\item{mu_dots}{the ratios of distance of generic order.}
}
\value{
the value of the log-likelihood evaluated in \code{par}.
}
\description{
Log-likelihood generated from the generic ratio of two NN's distances
}
\examples{
\dontrun{
loglikelihood_gride(d_value,n1=1,n2=2,mudotsample)
}

}
