% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gride.R
\name{mle_gride}
\alias{mle_gride}
\title{Maximum Likelihood Estimator (MLE) for the ID using generic ratios}
\usage{
mle_gride(X, n1 = 1, n2 = 2, nsim = 2000, conf_lev = 0.95)
}
\arguments{
\item{X}{a \code{n} \eqn{times} \code{D} data matrix.}

\item{n1}{order of the NN according to the first distance is computed.}

\item{n2}{order of the NN according to the second distance is computed.}

\item{nsim}{the number of bootstrap samples to consider.}

\item{conf_lev}{the confidence level for the computation of the confidence interval.}
}
\value{
the MLE estimate obtained via numeric optimization along with the bootstrap confidence interval.
}
\description{
Maximum Likelihood Estimator (MLE) for the ID using generic ratios
}
\examples{
\dontrun{
mle_gride(dataset,n1=1,n2=2)
}

}
