% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gride.R
\name{mle_gride_opt}
\alias{mle_gride_opt}
\title{Maximum Likelihood Estimator (MLE) for the ID using generic ratios}
\usage{
mle_gride_opt(X, n1 = 2, n2 = 2)
}
\arguments{
\item{X}{a \code{n} \eqn{times} \code{D} data matrix.}

\item{n1}{order of the NN according to the first distance is computed.}

\item{n2}{order of the NN according to the second distance is computed.}
}
\value{
the Gride MLE estimate obtained via numeric optimization.
}
\description{
Maximum Likelihood Estimator (MLE) for the ID using generic ratios
}
\examples{
\dontrun{
mle_gride_opt(dataset,n1=1,n2=2)
}

}
