% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoNN.R
\name{mle_twonn}
\alias{mle_twonn}
\title{Maximum Likelihood Estimator for the TWO-NN model}
\usage{
mle_twonn(
  X = NULL,
  unbiased = T,
  conf_lev = 0.95,
  trimmed = F,
  alpha_trimmed = 0.01,
  DM = NULL
)
}
\arguments{
\item{X}{a \code{n} \eqn{times} \code{D} data matrix.}

\item{unbiased}{logical, if \code{TRUE} the MLE is corrected to ensure unbiasedness.}

\item{conf_lev}{the confidence level for the computation of the confidence interval.}

\item{trimmed}{logical, if \code{TRUE}, the trimming of the most extreme observations is performed.}

\item{alpha_trimmed}{proportion of trimmed observations.}

\item{DM}{a distance matrix describing the nearest neighbors structure. Default is \code{NULL}.}
}
\value{
The MLE for the intrinsic dimension of the dataset \code{X} with its confidence interval.
}
\description{
Maximum Likelihood Estimator for the TWO-NN model
}
