% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Norm_Constant_Z_l2}
\alias{Norm_Constant_Z_l2}
\title{Compute the normalizing constant for the}
\usage{
Norm_Constant_Z_l2(Nzi_l, N, csi, q)
}
\arguments{
\item{Nzi_l}{the number of observations assigned to each manifold}

\item{N}{total number of}

\item{csi}{the strenght of the homogenity enforcement}

\item{q}{the number of NN considered for each observation}
}
\description{
Compute the normalizing constant for the
}
