% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gride.R
\name{plateau}
\alias{plateau}
\title{Plateau function}
\usage{
plateau(X, n.max, by = 2)
}
\arguments{
\item{X}{a \code{n} \eqn{times} \code{D} data matrix.}

\item{n.max}{integer,}

\item{by}{integer, step}
}
\value{
a matrix, with id estimates and average distances
}
\description{
Plateau function
}
\examples{
\dontrun{
plateau(dataset,n.max=100,by=2)
}
}
