% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gride.R
\name{posterior_sampler_gride}
\alias{posterior_sampler_gride}
\title{Posterior Metropolis Hasting sampler for the Grimes model}
\usage{
posterior_sampler_gride(
  X,
  DM = NULL,
  n1 = 1,
  n2 = 2,
  nsim = 5000,
  burn_in = 2000,
  sigma = 0.5,
  start_d = NULL,
  a_d = 1,
  b_d = 1
)
}
\arguments{
\item{X}{a \code{n} \eqn{times} \code{D} data matrix.}

\item{DM}{a distance matrix describing the nearest neighbors structure. Default is \code{NULL}.}

\item{n1}{order of the NN according to the first distance is computed.}

\item{n2}{order of the NN according to the second distance is computed.}

\item{nsim}{the number of MCMC iterations to collect in the sample.}

\item{burn_in}{the number of iterations to discard from the MCMC sample.}

\item{sigma}{standard deviation of the Gaussian proposal used in the MH step.}

\item{start_d}{initial value for the MCMC chain. If \eqn{NULL}, the MLE is used.}

\item{a_d}{shape parameter of the Gamma prior distribution for \eqn{d}.}

\item{b_d}{rate parameter of the Gamma prior distribution for \eqn{d}.}
}
\value{
a list with the coda MCMC sample and the traceplot.
}
\description{
Posterior Metropolis Hasting sampler for the Grimes model
}
\examples{
\dontrun{
posterior_sampler_gride(dataset, n1 = 1, n2 = 2)
}
}
