% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gride_bayes.R
\name{print.gride_bayes}
\alias{print.gride_bayes}
\title{Print \code{Gride} Bayes object}
\usage{
\method{print}{gride_bayes}(x, ...)
}
\arguments{
\item{x}{object of class \code{gride_bayes()}, obtained from the function
\code{gride_bayes()}.}

\item{...}{ignored.}
}
\value{
the function prints a summary of the Bayesian Gride to console.
}
\description{
Print \code{Gride} Bayes object
}
