% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gride_evolution.R
\name{print.gride_evolution}
\alias{print.gride_evolution}
\title{Print \code{Gride} evolution object}
\usage{
\method{print}{gride_evolution}(x, ...)
}
\arguments{
\item{x}{object of class \code{gride_evolution}, obtained from the function
\code{gride_evolution()}.}

\item{...}{ignored.}
}
\value{
the function prints a summary of the Gride evolution to
console.
}
\description{
Print \code{Gride} evolution object
}
