% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gride_mle.R
\name{print.gride_mle}
\alias{print.gride_mle}
\title{Print \code{Gride} MLE object}
\usage{
\method{print}{gride_mle}(x, ...)
}
\arguments{
\item{x}{object of class \code{gride_mle}, obtained from the function
\code{gride_mle()}.}

\item{...}{ignored.}
}
\value{
the function prints a summary of the Gride estimated via maximum
likelihood to console.
}
\description{
Print \code{Gride} MLE object
}
