% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidalgo_gibbs.R
\name{print.Hidalgo}
\alias{print.Hidalgo}
\title{Print the Hidalgo object}
\usage{
\method{print}{Hidalgo}(x, ...)
}
\arguments{
\item{x}{an object of class \code{Hidalgo}, obtained from the function
\code{Hidalgo()}.}

\item{...}{other arguments passed to specific methods.}
}
\value{
the function prints a summary of the Hidalgo MCMC run to console.
}
\description{
Print the Hidalgo object
}
