% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidalgo_aux.R
\name{print.hidalgo_psm}
\alias{print.hidalgo_psm}
\title{Print the summary of the clustering solution}
\usage{
\method{print}{hidalgo_psm}(x, ...)
}
\arguments{
\item{x}{object of class \code{hidalgo_psm}, obtained from the function
\code{psm_and_cluster()}.}

\item{...}{ignored.}
}
\value{
the function prints a summary of the clustering solution to console.
}
\description{
Print the summary of the clustering solution
}
