% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mus.R
\name{print.mus}
\alias{print.mus}
\title{Print the ratio statistics output}
\usage{
\method{print}{mus}(x, ...)
}
\arguments{
\item{x}{object of class \code{mus}, obtained from the
function \code{compute_mus()}.}

\item{...}{ignored.}
}
\value{
the function prints a summary of the computed ratio statistics to
console.
}
\description{
Print the ratio statistics output
}
