% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twonn_bayes.R
\name{print.twonn_bayes}
\alias{print.twonn_bayes}
\title{Print \code{TWO-NN} Bayes object}
\usage{
\method{print}{twonn_bayes}(x, ...)
}
\arguments{
\item{x}{object of class \code{twonn_bayes}, obtained from the function
\code{twonn_bayes()}.}

\item{...}{ignored.}
}
\value{
the function prints a summary of the Bayesian TWO-NN to console.
}
\description{
Print \code{TWO-NN} Bayes object
}
