% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twonn_dec_steps.R
\name{print.twonn_dec_by}
\alias{print.twonn_dec_by}
\title{Print \code{TWO-NN} evolution object decimated via halving steps}
\usage{
\method{print}{twonn_dec_by}(x, ...)
}
\arguments{
\item{x}{object of class \code{twonn_dec_prop}, obtained from the function
\code{twonn_dec_prop()}.}

\item{...}{ignored.}
}
\value{
the function prints a summary of the decimated TWO-NN to console.
}
\description{
Print \code{TWO-NN} evolution object decimated via halving steps
}
