% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twonn_linfit.R
\name{print.twonn_linfit}
\alias{print.twonn_linfit}
\title{Print TWO-NN Least Squares output}
\usage{
\method{print}{twonn_linfit}(x, ...)
}
\arguments{
\item{x}{object of class \code{twonn_linfit}, obtained from the function
\code{twonn_linfit()}.}

\item{...}{ignored.}
}
\value{
the function prints a summary of the TWO-NN estimated via least
squares to console.
}
\description{
Print TWO-NN Least Squares output
}
