% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twonn_mle.R
\name{print.twonn_mle}
\alias{print.twonn_mle}
\title{Print \code{TWO-NN} MLE output}
\usage{
\method{print}{twonn_mle}(x, ...)
}
\arguments{
\item{x}{object of class \code{twonn_mle}, obtained from the function
\code{twonn_mle()}.}

\item{...}{ignored.}
}
\value{
the function prints a summary of the TWO-NN estimated via
maximum likelihood to console.
}
\description{
Print \code{TWO-NN} MLE output
}
