% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PSM}
\alias{PSM}
\title{Efficient computation of the Posterior Coclustering Matrix}
\usage{
PSM(inds)
}
\arguments{
\item{inds}{a matrix containing the memebership labels for each observation (on the columns) in each MCMC iteration (on the rows)}
}
\description{
Efficient computation of the Posterior Coclustering Matrix
}
