% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoNN_Gride_Aux.R
\name{simulate_mudot}
\alias{simulate_mudot}
\title{Random sampler for the generalized ratio distribution of \eqn{\dot{\mu}}}
\usage{
simulate_mudot(nsim, n1 = 1, n2 = 2, d)
}
\arguments{
\item{nsim}{integer, the number of observations to generate.}

\item{n1}{order of the NN according to the first distance is computed.}

\item{n2}{order of the NN according to the second distance is computed.}

\item{d}{the value of the intrinsic dimension.}
}
\value{
a vector of random observations sampled from the generalized ratio distribution.
}
\description{
Random sampler for the generalized ratio distribution of \eqn{\dot{\mu}}
}
\examples{
simulate_mudot(100,3,5,2)
}
