% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_auxiliary.R
\name{Swissroll}
\alias{Swissroll}
\title{Generates a noise-free Swiss roll dataset}
\usage{
Swissroll(n)
}
\arguments{
\item{n}{number of observations contained in the output dataset.}
}
\value{
a three-dimensional \code{data.frame} containing the coordinates of
the points generated via the Swiss roll mapping.
}
\description{
The function creates a three-dimensional dataset with coordinates
following the Swiss roll mapping, transforming random uniform data points
sampled on the interval \code{(0,10)}.
}
\examples{
Data <- Swissroll(1000)

}
