% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_auxiliary.R
\name{Swissroll_maker}
\alias{Swissroll_maker}
\title{Generates a noise-free Swiss Roll dataset}
\usage{
Swissroll_maker(N)
}
\arguments{
\item{N}{number of observations contained in the output dataset}
}
\value{
a three-dimensional Swiss Roll dataset
}
\description{
Generates a noise-free Swiss Roll dataset
}
\examples{
Swissroll_maker(1000)
}
