% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twonn.R, R/twonn_bayes.R, R/twonn_linfit.R,
%   R/twonn_mle.R
\name{twonn}
\alias{twonn}
\alias{print.twonn_bayes}
\alias{summary.twonn_bayes}
\alias{print.summary.twonn_bayes}
\alias{plot.twonn_bayes}
\alias{print.twonn_linfit}
\alias{summary.twonn_linfit}
\alias{print.summary.twonn_linfit}
\alias{plot.twonn_linfit}
\alias{print.twonn_mle}
\alias{summary.twonn_mle}
\alias{print.summary.twonn_mle}
\alias{plot.twonn_mle}
\title{\code{TWO-NN} estimator}
\usage{
twonn(
  X = NULL,
  dist_mat = NULL,
  mus = NULL,
  method = c("mle", "linfit", "bayes"),
  alpha = 0.95,
  c_trimmed = 0.01,
  unbiased = TRUE,
  a_d = 0.001,
  b_d = 0.001,
  ...
)

\method{print}{twonn_bayes}(x, ...)

\method{summary}{twonn_bayes}(object, ...)

\method{print}{summary.twonn_bayes}(x, ...)

\method{plot}{twonn_bayes}(x, plot_low = 0.001, plot_upp = NULL, by = 0.05, ...)

\method{print}{twonn_linfit}(x, ...)

\method{summary}{twonn_linfit}(object, ...)

\method{print}{summary.twonn_linfit}(x, ...)

\method{plot}{twonn_linfit}(x, ...)

\method{print}{twonn_mle}(x, ...)

\method{summary}{twonn_mle}(object, ...)

\method{print}{summary.twonn_mle}(x, ...)

\method{plot}{twonn_mle}(x, ...)
}
\arguments{
\item{X}{data matrix with \code{n} observations and \code{D} variables.}

\item{dist_mat}{distance matrix computed between the \code{n} observations.}

\item{mus}{vector of second to first NN distance ratios.}

\item{method}{chosen estimation method. It can be
\describe{
      \item{\code{"mle"}}{for maximum likelihood estimator;}
      \item{\code{"linfit"}}{for estimation via the least squares approach;}
      \item{\code{"bayes"}}{for estimation with the Bayesian approach.}
             }}

\item{alpha}{the confidence level (for \code{mle} and least squares fit) or
posterior probability in the credible interval (\code{bayes}).}

\item{c_trimmed}{the proportion of trimmed observations.}

\item{unbiased}{logical, applicable when \code{method = "mle"}.
If \code{TRUE}, the MLE is corrected to ensure unbiasedness.}

\item{a_d}{shape parameter of the Gamma prior on the parameter \code{d},
applicable when \code{method = "bayes"}.}

\item{b_d}{rate parameter of the Gamma prior on the parameter \code{d},
applicable when \code{method = "bayes"}.}

\item{...}{ignored.}

\item{x}{object of class \code{twonn_mle}, the output of the
\code{twonn} function when \code{method = "mle"}.}

\item{object}{object of class \code{twonn_mle}, obtained from the function
\code{twonn_mle()}.}

\item{plot_low}{lower bound of the interval on which the posterior density
is plotted.}

\item{plot_upp}{upper bound of the interval on which the posterior density
is plotted.}

\item{by}{step-size at which the sequence spanning the interval is
incremented.}
}
\value{
list characterized by a class type that depends on the \code{method}
chosen. Regardless the \code{method}, the output list always contains the
object \code{est}, which provides the estimated intrinsic dimension along
with uncertainty quantification. The remaining objects varies with the
estimation method. In particular, if
\describe{
\item{\code{method = "mle"}}{the output reports the MLE and the relative
confidence interval;}
\item{\code{method = "linfit"}}{the output includes the \code{lm()} object
used for the computation;}
\item{\code{method = "bayes"}}{the output contains the {(1 + \code{alpha}) / 2}
and {(1 - \code{alpha}) / 2}
quantiles, mean, mode and median of the posterior distribution of \code{d}.}
}
}
\description{
The function can fit the two-nearest neighbor estimator within the maximum
likelihood and the Bayesian frameworks. Also, one can obtain the estimates
using least squares estimation, depending on the specification of the
argument \code{method}. This model has been originally presented in
\href{https://www.nature.com/articles/s41598-017-11873-y}{Facco et al., 2017}
. See also \href{https://arxiv.org/abs/2104.13832}{Denti et al., 2021+}
for more details.
}
\examples{
# dataset with 1000 observations and id = 2
X <- replicate(2,rnorm(1000))
twonn(X)
# dataset with 1000 observations and id = 3
Y <- replicate(3,runif(1000))
#  Bayesian and least squares estimate from distance matrix
dm <- as.matrix(dist(Y,method = "manhattan"))
twonn(dist_mat = dm,method = "bayes")
twonn(dist_mat = dm,method = "linfit")

}
\references{
Facco E, D'Errico M, Rodriguez A, Laio A (2017). "Estimating the intrinsic
dimension of datasets by a minimal neighborhood information."
Scientific Reports, 7(1), 1-8.
ISSN 20452322, doi: 10.1038/s41598-017-11873-y.

Denti F, Doimo D, Laio A, Mira A (2022+). "Distributional Results for
Model-Based Intrinsic Dimension Estimators."
arXiv preprint. 2104.13832, \url{https://arxiv.org/abs/2104.13832}.
}
