% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gride.R
\name{unn_posterior_gride_real}
\alias{unn_posterior_gride_real}
\title{Unnormalized log-posterior for Grimes estimator on the real scale}
\usage{
unn_posterior_gride_real(z, n1, n2, mu_dots, a_d = 1, b_d = 1)
}
\arguments{
\item{z}{the transformed ID value, according to \eqn{z=\log(d-1)}.}

\item{n1}{order of the NN according to the first distance is computed.}

\item{n2}{order of the NN according to the second distance is computed.}

\item{mu_dots}{the sample of distance ratios \eqn{\dot{\mu}}.}

\item{a_d}{shape parameter of the Gamma prior distribution for \eqn{d}.}

\item{b_d}{rate parameter of the Gamma prior distribution for \eqn{d}.}
}
\value{
the value of the unnormalized log-posterior distribution evaluated in \eqn{z}.
}
\description{
Unnormalized log-posterior for Grimes estimator on the real scale
}
\examples{
\dontrun{
unn_posterior_gride_real(z=3, n1=1, n2=2, mu_dots=mu_sample)
}

}
